Revert "Allowed EmpiricalBernsteinCopula not to be a proper copula"
This reverts commit 883a39c987ecaa916d1e6ca87e309f5656776875.

There was an infinite loop in EmpiricalBernsteinCopula::computeEntropy
when this distribution is a copula, because it does not call method
from CopulaImplementation.

In order to use CopulaImplementation methods, EmpiricalBernsteinCopula
must inherit from CopulaImplementation, and checks are added in all
methods to call ContinuousDistribution methods when CopulaImplementation
is not a copula.